content/svg/content/test/test_SVGLengthList-2.xhtml

<html xmlns="http://www.w3.org/1999/xhtml">
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=630760
-->
<head>
  <title>Tests specific to SVGLengthList</title>
  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=515116">Mozilla Bug 630760</a>
<p id="display"></p>
<div id="content" style="display:none;">
<svg id="svg" xmlns="http://www.w3.org/2000/svg" width="100" height="100">
  <text id="text">
    <set attributeName="x" to="10 20 30 40" begin="0" dur="indefinite"/>
  </text>
</svg>
</div>
<pre id="test">
<script class="testbody" type="text/javascript">
<![CDATA[

SimpleTest.waitForExplicitFinish();

/*
Check that the animVal list for 'x' on <text> gives the correct number of
items when examined for the FIRST time DURING animation.
*/

function run_tests()
{
  document.getElementById('svg').pauseAnimations();

  var text = document.getElementById("text");
  var list = text.x.animVal;

  is(list.numberOfItems, 4, 'Checking numberOfItems');

  SimpleTest.finish();
}

window.addEventListener("load", run_tests, false);

]]>
</script>
</pre>
</body>
</html>
